+++ tests/unit/helper.ts
@@ -14,6 +14,14 @@ export const wallet = {
   publicKey: "0xea526ba1710343d953461ff68641f1b7df5f23b9042ffa2d2a798d3adb3f3d6c",
 };
 
+export const Ed25519WalletTestObject = {
+  address: "0x28b829b524d7c24aa7fd8916573c814df766dae542f724e1cf8914536232c346",
+  mnemonic: "shoot island position soft burden budget tooth cruel issue economy destroy above",
+  path: "m/44'/637'/0'/0'/0'",
+  privateKey: "0x5d996aa76b3212142792d9130796cd2e11e3c445a93118c08414df4f66bc60ec",
+  publicKey: "0xea526ba1710343d953461ff68641f1b7df5f23b9042ffa2d2a798d3adb3f3d6c",
+};
+
 export const secp256k1WalletTestObject = {
   address: "0x4b4aa8759fcef40ba49e999409eb73a98252f44f6612a4de2b23bad5c37b15a6",
   mnemonic: "shoot island position soft burden budget tooth cruel issue economy destroy above",